* more compiler warning fixes for the lang...utils subdirectories   (based on amd64 recompile) * move externs in C files to ptexlib.h/&lua/luatex-api.h for now   (for most, this is not the final place yet) * move typedef const_lstring to stringpool.h   (used in pdfobj.c) * typecast patch for kpathsea/lib.h and cpascal.h   (probably temporary, PEB is likely to change xmalloc's prototype)
git-svn-id: https://serveur-svn.lri.fr/svn/modhel/luatex/trunk@3395 0b2b3880-5936-4365-a048-eb17d2e5a6bf

+++ source/texk/kpathsea/lib.h
@@ -213,8 +213,8 @@ extern KPSEDLL address xrealloc (address old_address, unsigned new_size);
 extern KPSEDLL address xcalloc (unsigned nelem, unsigned elsize);
 
 /* (Re)Allocate N items of type T using xmalloc/xrealloc.  */
-#define XTALLOC(n, t) ((t *) xmalloc ((n) * sizeof (t)))
+#define XTALLOC(n, t) ((t *) xmalloc ((unsigned)((n) * sizeof (t))))
 #define XTALLOC1(t) XTALLOC (1, t)
-#define XRETALLOC(addr, n, t) ((addr) = (t *) xrealloc (addr, (n) * sizeof(t)))
+#define XRETALLOC(addr, n, t) ((addr) = (t *) xrealloc (addr, (unsigned)((n) * sizeof(t))))
 
 #endif /* not KPATHSEA_LIB_H */
