+++ source/texk/web2c/luatexdir/ocp/ocp.c
@@ -130,7 +130,7 @@ void allocate_ocp_table(int ocp_number, int ocp_size)
         ocp_entries = 65536;
     }
     ocp_tables[ocp_number] =
-        (int *) xmalloc((unsigned) (1 + ocp_size) * sizeof(int));
+        (int *) xmalloc((unsigned) ((1 + (unsigned)ocp_size) * sizeof(int)));
     ocp_tables[ocp_number][0] = ocp_size;
     for (i = 1; i <= ocp_size; i++) {
         ocp_tables[ocp_number][i] = 0;
@@ -178,7 +178,7 @@ static void undump_ocp_table(int ocp_number)
     }
     undump_things(sizeword, 1);
     ocp_tables[ocp_number] =
-        (int *) xmalloc((unsigned) (1 + sizeword) * sizeof(int));
+        (int *) xmalloc((unsigned) ((1 + (unsigned)sizeword) * sizeof(int)));
     ocp_tables[ocp_number][0] = sizeword;
     undump_things(ocp_tables[ocp_number][1], sizeword);
 }
